refactor(charts): centralize chartEx parse in ooxml-common parse_chartex_part
CH5 PR3, the final edit of the shared-chart-parse pattern established in
#736/#742: the pptx `parse_chartex` (waterfall/treemap chartEx, cx: 2014
namespace) was the last per-crate ChartModel builder that walked the XML
and populated every field inline. Lift its body verbatim into
`ooxml_common::chart::parse_chartex_part(chartspace_root, resolver)`,
which returns a bare ChartModel; pptx `parse_chartex` becomes a thin
adapter that builds a PptxColorResolver and wraps the result in a
ChartElement graphic frame — mirroring `parse_legacy_chart` exactly.

Two mechanical substitutions during the lift, both byte-neutral for the
pptx resolver:
- colour resolution routes through `resolver.resolve_solid_fill` instead
  of pptx's local `parse_color_node(node, theme)` (PptxColorResolver's
  `resolve_solid_fill` calls exactly that);
- theme fallback faces come from `resolver.theme_major_font_latin()` /
  `theme_minor_font_latin()` instead of a direct `theme.get("+mj-lt")` /
  `"+mn-lt"` read (NIT-2 from PR1 review; PptxColorResolver returns the
  same map values).

The shared entry point is public so xlsx/docx can recognize chartEx parts
later (CH14) without another copy of the walk; wiring those callers is out
of scope here. The chart type stays the raw series `layoutId` string, so
adding sunburst/boxWhisker/funnel/histogram is a renderer concern (CH15),
not a parse change.

Byte-stability: parsing sample-2's chartEx (waterfall, one part) before
and after the lift yields SHA-identical ChartModel JSON (subtotalIndices
[0,5], per-label accent1/tx1 colours, barGapWidth 80, valAxisHidden, and
the +mj-lt/+mn-lt theme faces all preserved), and the full-presentation
JSON is likewise unchanged. Four direct `parse_chartex_part` unit tests
in ooxml-common pin the contract (waterfall full decoration set, treemap
leaf values with no subtotals, None without a series, category newline
flattening) so the byte-stability guarantee survives after the temporary
golden harness is gone.
